requirements.txt
Flask
coqui-tts
torch
boto3
python-dotenv